Seperate DayInfo Data Class from CalendarViewModel
